Add Coordinates module and refactor coordinate parsing: move parse_coordinates_automa to new Coordinates.jl, update usages in KML and tables modules for improved organization and clarity.
