+++ src/parsing.jl
@@ -1,3 +1,6 @@
+# include("Coordinates.jl")
+using .Coordinates: parse_coordinates_automa, coordinate_string
+
 # turn an XML.Node into a KMLFile by finding the <kml> element
 function KMLFile(doc::XML.AbstractXMLNode)
     i = findfirst(x -> x.tag == "kml", XML.children(doc))
@@ -144,12 +147,11 @@ end
 
 Base.write(o::Writable; kw...) = Base.write(stdout, o; kw...)
 
-#-----------------------------------------------------------------------------# XML.Node ←→ KMLElement
-typetag(T) = replace(string(T), r"([a-zA-Z]*\.)" => "", "_" => ":")
+# ─────────────────────────────────────────────────────────────────────────────
+# XML.Node ←→ KMLElement
+# ─────────────────────────────────────────────────────────────────────────────
 
-coordinate_string(x::Tuple) = join(x, ',')
-coordinate_string(x::StaticArraysCore.SVector) = join(x, ',')
-coordinate_string(x::Vector) = join(coordinate_string.(x), '\n')
+typetag(T) = replace(string(T), r"([a-zA-Z]*\.)" => "", "_" => ":")
 
 # KMLElement → Node
 Node(o::T) where {T<:Enums.AbstractKMLEnum} = XML.Element(typetag(T), o.value)
@@ -179,7 +181,9 @@ function Node(o::T) where {names,T<:KMLElement{names}}
     return XML.Node(XML.Element, tag, attributes, nothing, children)
 end
 
-#-----------------------------------------------------------------------------# object (or enum)
+# ─────────────────────────────────────────────────────────────────────────────
+# object() – main entry point for parsing XML nodes into KML objects
+# ─────────────────────────────────────────────────────────────────────────────
 
 const ENUM_NAMES_SET = Set(names(Enums; all = true))             # Get all names in Enums
 
@@ -334,119 +338,6 @@ function _object_slow(node::XML.AbstractXMLNode)
     return nothing
 end
 
-# ───  Coordinates parsing function using Automata.jl  ─────────────────────────────────────
-
-# ------------------------------------------------------------------
-# 1.  Build the regular expression that recognises a 2‑D or 3‑D
-#     coordinate list of the form  "x,y[,z][ … repeated … ]"
-# ------------------------------------------------------------------
-
-#? const coord_number_re = rep1(re"[0-9.+\-Ee]+") # Alternative
-const coord_number_re = rep1(re"[^\t\n\r ,]+")
-const coord_delim_re = rep1(re"[\t\n\r ,]+")
-
-const _coord_number_actions = onexit!(onenter!(coord_number_re, :mark), :number)
-
-const _coord_machine_pattern =
-    opt(coord_delim_re) * opt(_coord_number_actions * rep(coord_delim_re * _coord_number_actions)) * opt(coord_delim_re)
-
-const COORDINATE_MACHINE = compile(_coord_machine_pattern)
-
-# ------------------------------------------------------------------
-# 2.  Action table used by the FSM — created once, marked `const`
-# ------------------------------------------------------------------
-const PARSE_OPTIONS = Parsers.Options()#delim=nothing, quoted=false, stripwhitespace = false)
-const AUTOMA_COORD_ACTIONS = Dict{Symbol,Expr}(
-    # save the start position of a number
-    :mark => :(current_mark = p),
-
-    # convert the byte slice to Float64 and push!
-    :number => quote
-        #? println("Parsing: ", String(view(data_bytes, current_mark:p-1)))
-        push!(results_vector, Parsers.parse(Float64, view(data_bytes, current_mark:p-1), PARSE_OPTIONS))
-    end,
-)
-
-# ------------------------------------------------------------------
-# 3.  Generate the low‑level FSM driver exactly once
-#     and store it in the module’s global scope.
-# ------------------------------------------------------------------
-
-let ctx = CodeGenContext(vars = Variables(data = :data_bytes), generator = :goto)
-    eval(quote
-        function __core_automa_parser(data_bytes::AbstractVector{UInt8}, results_vector::Vector{Float64})
-            current_mark = 0
-
-            $(generate_init_code(ctx, COORDINATE_MACHINE))
-
-            p_end = sizeof(data_bytes)
-            p_eof = p_end
-
-            $(generate_exec_code(ctx, COORDINATE_MACHINE, AUTOMA_COORD_ACTIONS))
-
-            return cs          # final machine state
-        end
-    end)
-end
-
-# ------------------------------------------------------------------
-# 4.  High‑level convenience wrapper
-# ------------------------------------------------------------------
-
-"""
-    _parse_coordinates_automa(txt::AbstractString)
-
-Parse a KML/GeoRSS-style coordinate string and return a vector of
-`SVector{3,Float64}` (if the list length is divisible by 3) **or**
-`SVector{2,Float64}` (if divisible by 2).
-"""
-function _parse_coordinates_automa(txt::AbstractString)
-    parsed_floats = Float64[]
-    # sizehint!(parsed_floats, length(txt) ÷ 4)
-    # sizehint! does not bring any speedup here
-    final_state = __core_automa_parser(codeunits(txt), parsed_floats)
-
-    # --- basic FSM state checks -------------------------------------------------
-    if final_state < 0
-        error("Coordinate string is malformed (FSM error state $final_state).")
-    end
-    #? Check below if the FSM ended in a valid state dropping garbage at the end the string
-    #? This check is overly strict and is not done for now (May 2025)
-    # if final_state > 0 && !(final_state == COORDINATE_MACHINE.start_state && isempty(txt))
-    #     error("Coordinate string is incomplete or has trailing garbage (FSM state $final_state).")
-    # end
-
-    # --- assemble SVectors ------------------------------------------------------
-    len = length(parsed_floats)
-
-    if len == 0
-        return SVector{0,Float64}[]
-    elseif len % 3 == 0
-        n = len ÷ 3
-        result = Vector{SVector{3,Float64}}(undef, n)
-        @inbounds for i = 1:n
-            off = (i - 1) * 3
-            result[i] = SVector{3,Float64}(parsed_floats[off+1], parsed_floats[off+2], parsed_floats[off+3])
-        end
-        return result
-    elseif len % 2 == 0
-        n = len ÷ 2
-        result = Vector{SVector{2,Float64}}(undef, n)
-        @inbounds for i = 1:n
-            off = (i - 1) * 2
-            result[i] = SVector{2,Float64}(parsed_floats[off+1], parsed_floats[off+2])
-        end
-        return result
-    else # len is not 0 and not a multiple of 2 or 3
-        if !isempty(txt) && !all(isspace, txt)
-            snippet = first(txt, min(50, lastindex(txt)))
-            @warn "Parsed $len numbers from \"$snippet…\", which is not a multiple of 2 or 3. Returning empty coordinates." maxlog =
-                1
-        end
-        return SVector{0,Float64}[] # Return empty instead of erroring
-    end
-end
-
 # -----------------------------------------------------------------------------
 # Main add_element! function
 # -----------------------------------------------------------------------------
@@ -570,7 +461,7 @@ function _convert_and_set_simple_field!(
 
     # Dispatch based on field type
     if true_field_name === :coordinates || true_field_name === :gx_coord
-        parsed_coords_vec = _parse_coordinates_automa(processed_string_val)
+        parsed_coords_vec = parse_coordinates_automa(processed_string_val)
         final_val_to_set = nothing
         if isempty(parsed_coords_vec)
             final_val_to_set =
@@ -764,7 +655,7 @@ function _parse_and_set_scalar_field!(
                 isempty(processed_string_val) && Nothing <: ftype_original ? nothing :
                 non_nothing_ftype(processed_string_val)
         elseif true_field_name === :coordinates || true_field_name === :gx_coord
-            parsed_coords_vec = _parse_coordinates_automa(processed_string_val)
+            parsed_coords_vec = parse_coordinates_automa(processed_string_val)
             if isempty(parsed_coords_vec)
                 final_val_to_set =
                     Nothing <: ftype_original ? nothing :
@@ -925,7 +816,7 @@ function autosetfield!(o::Union{Object,KMLElement}, sym::Symbol, txt::String)
     elseif ftype <: Enums.AbstractKMLEnum
         ftype(txt)
     elseif fname === :coordinates
-        vec = _parse_coordinates_automa(txt)
+        vec = parse_coordinates_automa(txt)
         val = (ftype <: Union{Nothing,Tuple}) ? first(vec) : vec
     else
         txt   # last‑ditch: store the raw string
